requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.10
bs4==0.0.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
datasets==2.14.4
dill==0.3.7
ekphrasis==0.5.4
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==3.1.0
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
NRCLex==3.0.0
numpy==1.25.2
pandas==2.0.3
pathy==0.10.2
patsy==0.5.3
Pillow==10.0.0
plotly==5.16.1
pyarrow==13.0.0
pydantic==2.3.0
pydantic_core==2.6.3
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-posthocs==0.7.0
scipy==1.11.2
seaborn==0.12.2
sklearn==0.0.post7
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
statsmodels==0.14.0
sympy==1.12
tenacity==8.2.3
termcolor==2.3.0
textblob==0.17.1
thinc==8.1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.32.0
typer==0.9.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
user-agent==0.1.10
wasabi==1.1.2
xmltodict==0.13.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
ipywidgets
